Implement backend suggestion

+++ full-stack-engineer/submissions/shaphan/backend/src/controllers/reviews.controller.ts
@@ -1,7 +1,7 @@
 import { Request, Response } from 'express';
 import { validationResult } from 'express-validator';
 import {
-  createReview, deleteReview, ReviewInput, updateReview,
+  createReview, deleteReview, findReviewById, findUserReview, ReviewInput, updateReview,
 } from '../services/reviews.service';
 
 const addReview = async (
@@ -15,12 +15,18 @@ const addReview = async (
   }
 
   const { user } = res.locals;
+  if (await findUserReview(user.id, req.body.productId)) {
+    return res.status(400).json({
+      success: false,
+      message: 'You have already reviewed this product',
+    });
+  }
   const review = await createReview(req.body, user.id);
 
   return res.status(201).json({
     success: true,
     message: 'Review created successfully',
-    data: {
+    review: {
       ...review,
     },
   });
@@ -35,6 +41,15 @@ const editReview = async (
   const errors = validationResult(req);
   if (!errors.isEmpty()) {
     return res.status(400).json({ errors: errors.array(), success: false });
+  }
+
+  const { user } = res.locals;
+
+  if (!await findUserReview(user.id, req.body.productId)) {
+    return res.status(400).json({
+      success: false,
+      message: 'This review is not yours, you can\'t edit it.',
+    });
   }
 
   const review = await updateReview(params.id, req.body);
@@ -56,6 +71,16 @@ const removeReview = async (
   const errors = validationResult(req);
   if (!errors.isEmpty()) {
     return res.status(400).json({ errors: errors.array(), success: false });
+  }
+
+  const { user } = res.locals;
+  const rev = await findReviewById(params.id);
+
+  if (Number(rev?.userId) !== Number(user.id)) {
+    return res.status(400).json({
+      success: false,
+      message: 'This review is not yours, you can\'t delete it.',
+    });
   }
 
   const review = await deleteReview(params.id);
